rename to match command name
