+++ src/commands/init.js
