+++ src/main.js
@@ -1,7 +1,7 @@
 #!/usr/bin/env node
 import { Argument, Command, InvalidArgumentError } from 'commander';
 import { authAction } from './commands/auth.js';
-import { initializeAction } from './commands/initialize.js';
+import { initializeAction } from './commands/init.js';
 import { solveAction } from './commands/solve.js';
 import { statsAction } from './commands/stats.js';
 import { submitAction } from './commands/submit.js';
